[TST] Consolidate the dependence suite, and fix a dof bug it exposed
The suite swept every estimator over behaviour that lives entirely in
IBMAEstimator, so the same base-class lines were re-fitted eight or nine times
for one assertion each.

Split the parametrization by what is actually under test. Behaviour that cannot
vary by subclass now runs over three representative estimators, one per family;
behaviour that genuinely differs per class still sweeps all nine.

Merged:
- three tests of `_pymare_weighting_kwargs`, a four-line method, into one table
  covering scheme x whether groups exist
- `corr_matrix_built` / `corr_matrix_not_built` / `groups_are_found`, which
  asserted the same two facts across 17 fits, into one sweep of 9 that checks
  the correlation is present exactly when `_requires_corr_matrix` says
- `produces_usable_maps` / `with_liberal_mask` / `dof_map_is_float`, three
  sweeps re-fitting for one assertion each, into one parametrized over
  estimator x masking strategy
- the two regression-dof tests, which fitted the same estimators twice to assert
  complementary halves of one claim
- the two PermutedOLS block tests into one table over `groupby`

Dropped `groupby_false_skips_correction`, now covered by the merged
no-dependence test.

143 tests, down from 176, in 50s rather than 82s. The reduction is in repeated
fits of identical code paths, not in coverage: the merged tests assert strictly
more per fit than the ones they replace.

The consolidation immediately earned itself. Sweeping the merged map test over
both masking strategies caught that the liberal-mask path pre-allocated
`dof_map` as int32 and assigned the new float degrees of freedom into it,
truncating them back and reinstating the INT_MIN out-of-mask values the previous
commit had just removed. Eight allocations fixed. No test covered that path
before, because the dtype assertion and the liberal-mask fit lived in different
tests.

+++ nimare/meta/ibma.py
@@ -561,7 +561,7 @@ class Fishers(IBMAEstimator):
             n_voxels = self.inputs_["z_maps"].shape[1]
             z_map = np.zeros(n_voxels, dtype=float)
             p_map = np.zeros(n_voxels, dtype=float)
-            dof_map = np.zeros(n_voxels, dtype=np.int32)
+            dof_map = np.zeros(n_voxels, dtype=float)
             for bag in self.inputs_["data_bags"]["z_maps"]:
                 (
                     z_map[bag["voxel_mask"]],
@@ -771,7 +771,7 @@ class Stouffers(IBMAEstimator):
             n_voxels = self.inputs_["z_maps"].shape[1]
             z_map = np.zeros(n_voxels, dtype=float)
             p_map = np.zeros(n_voxels, dtype=float)
-            dof_map = np.zeros(n_voxels, dtype=np.int32)
+            dof_map = np.zeros(n_voxels, dtype=float)
             for bag in self.inputs_["data_bags"]["z_maps"]:
                 (
                     z_map[bag["voxel_mask"]],
@@ -953,7 +953,7 @@ class WeightedLeastSquares(_PyMARERegressionEstimator):
             n_voxels = self.inputs_["beta_maps"].shape[1]
 
             z_map, p_map, est_map, se_map = [np.zeros(n_voxels, dtype=float) for _ in range(4)]
-            dof_map = np.zeros(n_voxels, dtype=np.int32)
+            dof_map = np.zeros(n_voxels, dtype=float)
 
             beta_bags = self.inputs_["data_bags"]["beta_maps"]
             varcope_bags = self.inputs_["data_bags"]["varcope_maps"]
@@ -1136,7 +1136,7 @@ class DerSimonianLaird(_PyMARERegressionEstimator):
             z_map, p_map, est_map, se_map, tau2_map = [
                 np.zeros(n_voxels, dtype=float) for _ in range(5)
             ]
-            dof_map = np.zeros(n_voxels, dtype=np.int32)
+            dof_map = np.zeros(n_voxels, dtype=float)
 
             beta_bags = self.inputs_["data_bags"]["beta_maps"]
             varcope_bags = self.inputs_["data_bags"]["varcope_maps"]
@@ -1324,7 +1324,7 @@ class Hedges(_PyMARERegressionEstimator):
             z_map, p_map, est_map, se_map, tau2_map = [
                 np.zeros(n_voxels, dtype=float) for _ in range(5)
             ]
-            dof_map = np.zeros(n_voxels, dtype=np.int32)
+            dof_map = np.zeros(n_voxels, dtype=float)
 
             beta_bags = self.inputs_["data_bags"]["beta_maps"]
             varcope_bags = self.inputs_["data_bags"]["varcope_maps"]
@@ -1528,7 +1528,7 @@ class SampleSizeBasedLikelihood(_PyMARERegressionEstimator):
             z_map, p_map, est_map, se_map, tau2_map, sigma2_map = [
                 np.zeros(n_voxels, dtype=float) for _ in range(6)
             ]
-            dof_map = np.zeros(n_voxels, dtype=np.int32)
+            dof_map = np.zeros(n_voxels, dtype=float)
 
             for bag in self.inputs_["data_bags"]["beta_maps"]:
                 (
@@ -1736,7 +1736,7 @@ class VarianceBasedLikelihood(_PyMARERegressionEstimator):
             z_map, p_map, est_map, se_map, tau2_map = [
                 np.zeros(n_voxels, dtype=float) for _ in range(5)
             ]
-            dof_map = np.zeros(n_voxels, dtype=np.int32)
+            dof_map = np.zeros(n_voxels, dtype=float)
 
             beta_bags = self.inputs_["data_bags"]["beta_maps"]
             varcope_bags = self.inputs_["data_bags"]["varcope_maps"]
@@ -2252,7 +2252,7 @@ class FixedEffectsHedges(_PyMARERegressionEstimator):
             n_voxels = self.inputs_["t_maps"].shape[1]
 
             z_map, p_map, est_map, se_map = [np.zeros(n_voxels, dtype=float) for _ in range(4)]
-            dof_map = np.zeros(n_voxels, dtype=np.int32)
+            dof_map = np.zeros(n_voxels, dtype=float)
 
             for bag in self.inputs_["data_bags"]["t_maps"]:
                 (
